article_fetcher.py
from newspaper import Article

def fetch_article_text(url: str) -> str:
    try:
        article = Article(url)
        article.download()
        article.parse()
        return article.text
    except Exception as e:
        print(f"Failed to fetch: {url} | {e}")
        return ""