requirements.txt
requests
beautifulsoup4
pandas
openpyxl
newspaper3k
groq
python-dotenv
lxml
lxml_html_clean
duckduckgo_search